Namespaced Filters & Dried up the code
Instead of repeating the networks over and over again, I think defining
them once would be best here. Now it's easier to add more in later on
(or remove some).

Also this helps to add filters on later ( I'll work on that very very
soon ).

This is not the best solution right now. Instead of using Variable
variables, we should use extract() on array in the template parts, but
that's something for another day.