Modified the rendering of networks and more
Networks now can be rendered by a custom function that user can define in functions.php on the theme side. Useful for Theme Developers.

Also Added the option to remove some options if they aren't supported, that way if the options aren't there, user won't be fooled later on. For example if the theme doesn't support Tooltips, but the user has set tooltips on.