Rename to Pure Social Icons
Renamed the visible Admin part to Pure Social Icons.
Everything else still is tipsy-social-icons.

Not sure whether to rename tipsy to pure in the source, as it would complicate merging back to Tipsy later on. I'll just leave this decision for later. Let future self deal with this. (you're welcome)